backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
libtorrent==2.0.9
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.12.2
urllib3==2.2.3
